LLM Flowchart Generation Mini-Project (Python)/requirements.txt
streamlit
mermaid-py
python-dotenv